ensure allowed_users is an array
When using node attributes, the returned result from directly accessing them is
an immuateable object that would throw errors. Convert the parent object to a
hash (wasn't sure what would happen calling to_hash on an array-like object) and
access the members that way. The result is a basic array, which can be
manipulated freely. It'll eventually be persisted via the later group resource,
so no need to call `node.save!` for the updated value.