Override settings from passwd to run as a restricted user
The default behavior for dovecot's lmtp is to setuid/setgid, which it can't do
if it's already running as a restricted user. Instead, override the return
values from the passwd lookup to do 2 things: move the mail directory to outside
of the user's home, and to change the destination user. Thus, lmtp will see it's
already the 'correct' user and won't try to setuid/setgid.